Encrypted account key storage
CB-436

Account keys are now stored in a single JSON file per key pair, which is named "keypair<index>.json".
The format corresponds to the "EncryptedAccountKeyPair" datatype.

Whenever encrypted account keys are needed to sign transactions or to produce proves, the encryption password is needed.
For command-line usage, the password will be queried, for API requests, the password has to be sent as part of the request.
Therefore the middleware API changed (see diff) and the components depending on the API have to be update accordingly.

Changes to CLI commands:
- Whenever account keys are provided via the command line (--keys), they are now expected to be encrypted and have to be provided in a file.
- The "config key add" command has been replaced by "config add-keys" and now also expects encrypted account keys
in a JSON file.
- The command "baker set-account" now expects a reference to an account present in the config (instead of a file with keys).

About some new modules:
- Module "Account" now defines key pair and key map types for unencrypted and encrypted keys as well as encryption/decryption functions.
- Module "Export" replaces "WalletExport" and will contain formats and decoding/encoding functions for all supported exports.