fix(delegation-server): error handling done in a more reasonable way
Several changes have been made:

1. Given a bit more descriptive names to some variables and types.

2. Each transition uses `handle` instead of `onException` so
exceptions are catched on each transition and the state is restored to
the previous one. As the state is locked during that transition, it
should be safe to do this.

3. Finalization listener will retry the transaction unless we get a
rejected transaction or a transaction that is finalized in several
blocks (which indicates a critical error). For rejected transactions,
it depends whether we provide a rollback or not, the transaction will
be retried or the state will be rolled back. Specifically when getting
a rejected delegation we rollback and when getting a rejected
undelegation we retry the transaction.

4. The error calls when a baker is missing in some state now continue
instead of throwing an exception using a state that doesn't have said
baker.