Set RPATH so that all executables and tests can be run without setting LD_LIBRARY_PATH.
Do so only for non-static builds on Linux, the behaviour on other platforms is unchanged.
The embedded path is the absolute one, which is what we really want since it
makes it possible to run the executable from any directory. This is the correct
thing for local use, for distribution of the binary and global installation we
should use the static binaries instead.