Fix most tests, and comment out those which rely on JSON layout.
The type of the credential is now returned together with the credential.

Tests generally should not rely on the details of the JSON return value (e.g.
amount of spaces, order of keys), since that is an implementation detail of the
JSON serializer that is not documented, and thus should not be relied upon.