Adding nix flake support
This branch should not be merged because Concordium does not intend to
commit to nix as a developement tool. It is only present since it
at least helps me, and is pushed as it _may_ help others.

Furthermore, this flake is rather shoddy, and does not even meet my own
standards in order to be merged to main. It does work however and allows you
to develop on this project with pinned dependencies that does not pollute
your system environment.